Commit early Mac functions
These changes make the application open on Mac. It doesn’t do much more.

+++ PROGRAM/Scripts/lockdown.py
@@ -1,4 +1,8 @@
-import win32gui, win32con, sys, winreg, subprocess
+## Import the required modules
+import sys, subprocess
+# Import the required modules based on platform
+if sys.platform == 'win32':
+    import win32gui, win32con, winreg
 
 
 # Set Registry Values for disabling Task Manager
@@ -8,38 +12,46 @@ value: int = 1
 
 
 def firstlockdown():
-    # Check if we are running in a compiled pyinstaller .exe, if not, don't run lockdown code
-    if getattr(sys, 'frozen', False):
-        # Find the "Pomoimprove" window, set as "hwnd"
-        hwnd = win32gui.FindWindowEx(None, None, None, "Pomoimprove")
-        # Put "Pomoimprove" window on top
-        win32gui.SetWindowPos(
-            hwnd,
-            win32con.HWND_TOPMOST,
-            0,
-            100,
-            200,
-            200,
-            win32con.SWP_NOMOVE | win32con.SWP_NOSIZE,
-        )
-
-        # Attempt to kill Explorer and Task Manager
-        try:
-            DETACHED_PROCESS = 0x00000008
-            subprocess.call('taskkill /f /im explorer.exe', creationflags=DETACHED_PROCESS)
-            subprocess.call('taskkill /f /im taskmgr.exe', creationflags=DETACHED_PROCESS)
-        except:
-            pass
-
-        # Edit Registry to disable Task Manager
-        reg_key = winreg.CreateKeyEx(winreg.HKEY_CURRENT_USER, registry_path, 0, winreg.KEY_SET_VALUE)
-        winreg.SetValueEx(reg_key, registry_name, 0, winreg.REG_DWORD, value)
-        winreg.CloseKey(reg_key)
+    # Check if we are running on macOS, if so, don't run anything.
+    if sys.platform == 'darwin':
+        print('Coming Soon.')
+    elif sys.platform == 'win32':
+        # Check if we are running in a compiled pyinstaller .exe, if not, don't run lockdown code
+        if getattr(sys, 'frozen', False):
+            # Find the "Pomoimprove" window, set as "hwnd"
+            hwnd = win32gui.FindWindowEx(None, None, None, "Pomoimprove")
+            # Put "Pomoimprove" window on top
+            win32gui.SetWindowPos(
+                hwnd,
+                win32con.HWND_TOPMOST,
+                0,
+                100,
+                200,
+                200,
+                win32con.SWP_NOMOVE | win32con.SWP_NOSIZE,
+            )
+    
+            # Attempt to kill Explorer and Task Manager
+            try:
+                DETACHED_PROCESS = 0x00000008
+                subprocess.call('taskkill /f /im explorer.exe', creationflags=DETACHED_PROCESS)
+                subprocess.call('taskkill /f /im taskmgr.exe', creationflags=DETACHED_PROCESS)
+            except:
+                pass
+    
+            # Edit Registry to disable Task Manager
+            reg_key = winreg.CreateKeyEx(winreg.HKEY_CURRENT_USER, registry_path, 0, winreg.KEY_SET_VALUE)
+            winreg.SetValueEx(reg_key, registry_name, 0, winreg.REG_DWORD, value)
+            winreg.CloseKey(reg_key)
 
 
 def secondlockdown():
-    # Check if we are running in a compiled pyinstaller .exe, if not, don't run lockdown code
-    if getattr(sys, 'frozen', False):
-        # This code will close all windows, except for the main Pomoimprove window
-        DETACHED_PROCESS = 0x00000008
-        subprocess.call('powershell -EncodedCommand RwBlAHQALQBQAHIAbwBjAGUAcwBzACAAfAAgAFcAaABlAHIAZQAtAE8AYgBqAGUAYwB0ACAAewAoACQAXwAuAE0AYQBpAG4AVwBpAG4AZABvAHcAVABpAHQAbABlACAALQBuAGUAIAAiACIAKQAgAC0AYQBuAGQAIAAoACQAXwAuAE0AYQBpAG4AVwBpAG4AZABvAHcAVABpAHQAbABlACAALQBuAGUAIAAiAFAAbwBtAG8AaQBtAHAAcgBvAHYAZQAiACkAfQAgAHwAIABzAHQAbwBwAC0AcAByAG8AYwBlAHMAcwA=', creationflags=DETACHED_PROCESS)
+    # Check if we are running on macOS, if so, don't run anything.
+    if sys.platform == 'darwin':
+        print('Coming Soon.')
+    elif sys.platform == 'win32':
+        # Check if we are running in a compiled pyinstaller .exe, if not, don't run lockdown code
+        if getattr(sys, 'frozen', False):
+            # This code will close all windows, except for the main Pomoimprove window
+            DETACHED_PROCESS = 0x00000008
+            subprocess.call('powershell -EncodedCommand RwBlAHQALQBQAHIAbwBjAGUAcwBzACAAfAAgAFcAaABlAHIAZQAtAE8AYgBqAGUAYwB0ACAAewAoACQAXwAuAE0AYQBpAG4AVwBpAG4AZABvAHcAVABpAHQAbABlACAALQBuAGUAIAAiACIAKQAgAC0AYQBuAGQAIAAoACQAXwAuAE0AYQBpAG4AVwBpAG4AZABvAHcAVABpAHQAbABlACAALQBuAGUAIAAiAFAAbwBtAG8AaQBtAHAAcgBvAHYAZQAiACkAfQAgAHwAIABzAHQAbwBwAC0AcAByAG8AYwBlAHMAcwA=', creationflags=DETACHED_PROCESS)
